# contracts/openapi.json
{
  "components": {
    "schemas": {
      "Block": {
        "additionalProperties": false,
        "properties": {
          "bbox": {
            "examples": [
              [
                0.1,
                0.2,
                0.9,
                0.3
              ]
            ],
            "items": {
              "type": "number"
            },
            "maxItems": 4,
            "minItems": 4,
            "title": "Bbox",
            "type": "array"
          },
          "block_id": {
            "examples": [
              "p00001_b001"
            ],
            "pattern": "^p[0-9]{5}_b[0-9]{3}$",
            "title": "Block Id",
            "type": "string"
          },
          "confidence": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              null
            ],
            "title": "Confidence"
          },
          "image_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              null
            ],
            "title": "Image Key"
          },
          "level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              null
            ],
            "title": "Level"
          },
          "order": {
            "examples": [
              1
            ],
            "title": "Order",
            "type": "integer"
          },
          "page_id": {
            "examples": [
              "p00001"
            ],
            "pattern": "^p[0-9]{5}$",
            "title": "Page Id",
            "type": "string"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "Photosynthesis is the process by which green plants make food."
            ],
            "title": "Text"
          },
          "type": {
            "$ref": "#/components/schemas/BlockType",
            "examples": [
              "paragraph"
            ]
          }
        },
        "required": [
          "block_id",
          "page_id",
          "type",
          "bbox",
          "order"
        ],
        "title": "Block",
        "type": "object"
      },
      "BlockType": {
        "enum": [
          "heading",
          "paragraph",
          "list_item",
          "figure",
          "caption",
          "table",
          "formula",
          "header",
          "footer",
          "page_number"
        ],
        "title": "BlockType",
        "type": "string"
      },
      "CanonicalChapter": {
        "additionalProperties": false,
        "properties": {
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/Block"
            },
            "title": "Blocks",
            "type": "array"
          },
          "chapter_id": {
            "examples": [
              "6f1c2a4e-2b7d-4c1e-9a3f-1d2e3f4a5b6c"
            ],
            "format": "uuid",
            "title": "Chapter Id",
            "type": "string"
          },
          "content_hash": {
            "examples": [
              "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
            ],
            "pattern": "^[a-f0-9]{64}$",
            "title": "Content Hash",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "2026-09-14T10:00:00Z"
            ],
            "title": "Created At"
          },
          "language": {
            "examples": [
              "en"
            ],
            "title": "Language",
            "type": "string"
          },
          "ocr_engine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "pdf-text-layer"
            ],
            "title": "Ocr Engine"
          },
          "page_count": {
            "examples": [
              60
            ],
            "minimum": 1.0,
            "title": "Page Count",
            "type": "integer"
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/Page"
            },
            "title": "Pages",
            "type": "array"
          },
          "schema_version": {
            "const": 1,
            "examples": [
              1
            ],
            "title": "Schema Version",
            "type": "integer"
          },
          "source_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "jesc106.pdf"
            ],
            "title": "Source Filename"
          },
          "title": {
            "examples": [
              "Life Processes"
            ],
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "chapter_id",
          "content_hash",
          "title",
          "language",
          "page_count",
          "pages",
          "blocks"
        ],
        "title": "CanonicalChapter",
        "type": "object"
      },
      "CreateDocumentRequest": {
        "properties": {
          "filename": {
            "examples": [
              "jesc106.pdf"
            ],
            "title": "Filename",
            "type": "string"
          },
          "size_bytes": {
            "examples": [
              4200000
            ],
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "filename",
          "size_bytes"
        ],
        "title": "CreateDocumentRequest",
        "type": "object"
      },
      "CreateDocumentResponse": {
        "properties": {
          "document_id": {
            "examples": [
              "0b8e7c1a-5d2f-4e3b-8a9c-7d6e5f4a3b2c"
            ],
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "upload_url": {
            "examples": [
              "https://r2.example.com/uploads/0b8e...?X-Amz-Signature=..."
            ],
            "title": "Upload Url",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "upload_url"
        ],
        "title": "CreateDocumentResponse",
        "type": "object"
      },
      "DocumentError": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "examples": [
              "ENCRYPTED_PDF"
            ]
          },
          "message": {
            "examples": [
              "This PDF is password protected."
            ],
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "DocumentError",
        "type": "object"
      },
      "DocumentStatus": {
        "enum": [
          "uploaded",
          "processing",
          "ready",
          "failed"
        ],
        "title": "DocumentStatus",
        "type": "string"
      },
      "DocumentSummary": {
        "properties": {
          "chapter_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              null
            ],
            "title": "Chapter Id"
          },
          "created_at": {
            "examples": [
              "2026-09-14T10:00:00Z"
            ],
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "examples": [
              "0b8e7c1a-5d2f-4e3b-8a9c-7d6e5f4a3b2c"
            ],
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentError"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              null
            ]
          },
          "filename": {
            "examples": [
              "jesc106.pdf"
            ],
            "title": "Filename",
            "type": "string"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              60
            ],
            "title": "Page Count"
          },
          "progress": {
            "examples": [
              42
            ],
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus",
            "examples": [
              "processing"
            ]
          }
        },
        "required": [
          "document_id",
          "filename",
          "status",
          "progress",
          "created_at"
        ],
        "title": "DocumentSummary",
        "type": "object"
      },
      "ErrorCode": {
        "enum": [
          "SCAN_TOO_POOR",
          "ENCRYPTED_PDF",
          "TOO_MANY_PAGES",
          "CORRUPT_FILE"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IngestResponse": {
        "properties": {
          "job_id": {
            "examples": [
              "9a8b7c6d-5e4f-4a3b-9c2d-1e0f9a8b7c6d"
            ],
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "IngestResponse",
        "type": "object"
      },
      "Page": {
        "additionalProperties": false,
        "properties": {
          "has_text_layer": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              true
            ],
            "title": "Has Text Layer"
          },
          "height_pt": {
            "examples": [
              842.0
            ],
            "title": "Height Pt",
            "type": "number"
          },
          "image_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "examples": [
              "pages/3f9a.../p00001.webp"
            ],
            "title": "Image Key"
          },
          "index": {
            "examples": [
              0
            ],
            "minimum": 0.0,
            "title": "Index",
            "type": "integer"
          },
          "page_id": {
            "examples": [
              "p00001"
            ],
            "pattern": "^p[0-9]{5}$",
            "title": "Page Id",
            "type": "string"
          },
          "rotation": {
            "default": 0,
            "enum": [
              0,
              90,
              180,
              270
            ],
            "examples": [
              0
            ],
            "title": "Rotation",
            "type": "integer"
          },
          "width_pt": {
            "examples": [
              595.0
            ],
            "title": "Width Pt",
            "type": "number"
          }
        },
        "required": [
          "page_id",
          "index",
          "width_pt",
          "height_pt"
        ],
        "title": "Page",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "NotesApp API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/documents": {
      "get": {
        "operationId": "list_documents_v1_documents_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentSummary"
                  },
                  "title": "Response List Documents V1 Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Documents"
      },
      "post": {
        "operationId": "create_document_v1_documents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateDocumentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Document"
      }
    },
    "/v1/documents/{document_id}": {
      "get": {
        "operationId": "get_document_v1_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Document"
      }
    },
    "/v1/documents/{document_id}/chapter": {
      "get": {
        "operationId": "get_chapter_v1_documents__document_id__chapter_get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CanonicalChapter"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chapter"
      }
    },
    "/v1/documents/{document_id}/ingest": {
      "post": {
        "operationId": "ingest_document_v1_documents__document_id__ingest_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Document"
      }
    },
    "/v1/pages/{page_id}/image": {
      "get": {
        "operationId": "get_page_image_v1_pages__page_id__image_get",
        "parameters": [
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "title": "Page Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Redirect to a signed R2 URL, 15 min expiry"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Page Image"
      }
    }
  }
}